move fields classes into their own file
